Difficulty

+++ BlockSimSharp.Core/Configuration/Model/SimulationSettings.cs
@@ -4,5 +4,5 @@ namespace BlockSimSharp.Core.Configuration.Model;
 
 public sealed class SimulationSettings: ISettings
 {
-    public int LengthInSeconds { get; init; }
+    public float LengthInSeconds { get; init; }
 }
